Adds Proxy-Script
The proxy script allows loading of files out of directories not in the
document-root. This way includable files can be loaded from everywhere
on the filesystem as long as the containing folder has been allowed.

That way includable files from any module will not have to be copied to
a public directory but will be served directly out of their
module-folder